Fix undefined-var-template warning on Clang
Fixes a warning emitted by newer versions of Clang that warn about
static variables in templates that might be undefined. However, the
variables are indeed defined since otherwise the code would not link.
The warning is removed by making it clear to the compiler that the
variable is defined in some translation unit, which is achieved by
explicitly instantiating the template.