Fix LAPACK CMake and link problems on Linux
Fix problems with finding LAPACK and BLAS on Linux using CMake. The
created imported targets need to split the LAPACK_LIBRARIES variable.
For static MKL, cyclic dependencies are resolved using -Wl,--start-group
and -Wl,--end-group. This confuses the logic that constructs the
imported targets.

This problem is fixed by reverting to variables instead of imported
targets.